--- app/src/main/java/com/webkul/mobikul/odoo/ui/learning_anko/data/City.kt
@@ -1,10 +1,0 @@
-package com.webkul.mobikul.odoo.ui.learning_anko.data
-
-data class City(
-    val coordinates: Coordinates,
-    val country: String,
-    val id: Int,
-    val name: String,
-    val population: Int,
-    val timezone: Int
-)
