+++ app/src/main/java/com/webkul/mobikul/odoo/ui/learning_anko/data/models/City.kt
@@ -1,0 +1,10 @@
+package com.webkul.mobikul.odoo.ui.learning_anko.data.models
+
+data class City(
+	val coordinates: Coordinates,
+	val country: String,
+	val id: Int,
+	val name: String,
+	val population: Int,
+	val timezone: Int
+)
