+++ app/src/main/java/com/webkul/mobikul/odoo/ui/learning_anko/data/request/Request.kt
@@ -1,4 +1,4 @@
-package com.webkul.mobikul.odoo.ui.learning_anko.data
+package com.webkul.mobikul.odoo.ui.learning_anko.data.request
 
 import timber.log.Timber
 import java.net.URL
